Changes to prepare for the future:tm:
