+++ src/inject/dynamic-theme/css-collection.ts
@@ -1,18 +1,6 @@
 import {forEach} from '../../utils/array';
 import {loadAsDataURL} from '../../utils/network';
-import {getMatches} from '../../utils/text';
-
-function getShift(deep: number, isEndBracket: boolean) {
-    if (deep === 0) {
-        return '';
-    }
-    if (deep === 1) {
-        return isEndBracket ? ' '.repeat(4) : ' '.repeat(3);
-    }
-    if (deep > 1) {
-        return isEndBracket ? ' '.repeat(4).repeat(deep) : ' '.repeat(4).repeat(deep).substr(1); // All property's have by default already 1x ' '
-    }
-}
+import {getMatches, formatCSS} from '../../utils/text';
 
 const blobRegex = /url\(\"(blob\:.*?)\"\)/g;
 
@@ -24,32 +12,6 @@ async function replaceBlobs(text: string) {
     });
     const data = await Promise.all(promises);
     return text.replace(blobRegex, () => `url("${data.shift()}")`);
-}
-
-async function formatCSS(text: string) {
-    const CSS = (text
-        .replace(/(.*?){ }/g, '') // Removing Empty CSS Rules
-        .replace(/\s\s+/g, ' ') // Replacing multiple spaces to one
-        .replace(/\{/g,'{%--%') // {
-        .replace(/\}/g,'%--%}%--%') // }
-        .replace(/\;(?![^\(]*\))/g,';%--%') // ; and do not target between () mostly for url()
-        .replace(/%--%\s{0,}%--%/g,'%--%') // Remove %--% Without any characters between it to the next %--%
-        .split('%--%'));
-    let deep = 0;
-    const formatted = [];
-
-    for (let x = 0, len = CSS.length; x < len; x++) {
-        const line = CSS[x] + '\n';
-        if (line.match(/\{/)) { // {
-            formatted.push(getShift(deep++, false) + line);
-        } else if (line.match(/\}/)) { // }
-            formatted.push(getShift(--deep, true) + line);
-        } else { // CSS line
-            formatted.push(getShift(deep, false) + line);
-        }
-    }
-
-    return replaceBlobs(formatted.join(''));
 }
 
 export async function collectCSS() {
@@ -79,8 +41,9 @@ export async function collectCSS() {
     });
 
     if (modifiedCSS.length != 0) {
+        const formattedCSS = formatCSS(modifiedCSS.join('\n'));
         css.push('/* Modified CSS */');
-        css.push(await formatCSS(modifiedCSS.join('\n')));
+        css.push(await replaceBlobs(formattedCSS));
         css.push('');
     }
